automata: add new 'WhichCaptures' config
This is the first step in fixing a regression in memory usage. The
underlying problem is that regex-automata now natively supports
multi-pattern regexes *with* capturing support. Unfortunately though,
this overall doesn't work too well with the current design of the
PikeVM, because the amount of memory used is `len(captures) *
len(states)`. So basically, as the regex and number of captures
increases, the amount of memory used gets quite high.

This is new functionality that we hope to improve upon over time, so
it's not too big of a deal on its own. But it turns out this impacts
previous uses of RegexSet that have capture groups. The old
implementation just ignored these capture groups because they weren't
supported in a RegexSet, and thus there were no memory problems. But in
the new implementation, nothing tells it that it's okay to ignore the
capture groups. So it winds up allocating space for them even though the
RegexSet APIs don't provide any of that functionality.

So my plan to fix this is to introduce a new configuration knob for
controlling more granularly which capture states are compiled into the
NFA. Previously we only supported "all of them" or "none of them." This
commit adds a new (backwards compatible) knob that also permits "just
implicit groups." That is, one capture group per pattern. This hopefully
leads to less memory usage overall. (Well, it will certaintly be less,
but hopefully it's a big reduction.) We don't actually change anything
here. We just add a new `Config::which_captures` knob, implement the
existing `Config::captures` in terms of `Config::which_captures` and
deprecate `Config::captures`.

If this winds up not being sufficient, then we may need to adapt the
PikeVM to work without any capture groups at all and instead just report
which patterns match. Which is... probably fine?