automata: fix sparse DFA state validation in no-std
The verified set isn't tracked in no-std/no-alloc because it is probably
not worth doing (and bloating the size of the sparse DFA itself to store
the state IDs).

So when we deserialize a DFA without std enabled, the verified set of
states was always reporting `false`, and this now trips an error 100% of
the time in the new start state validation code.

We fix this by always reporting `true`, thus treating every possible
state ID as possibly valid on its own. Not great, but maintains the
status quo.