automata/meta: revert broadening of reverse suffix optimization
This reverts commit 8a8d599f9d2f2d78e9ad84e4084788c2d563afa5 and
includes a regression test, as well as a tweak to a log message.

Essentially, the broadening was improper. We have to be careful when
dealing with suffixes as opposed to prefixes. Namely, my logic
previously was that the broadening was okay because we were already
doing it for the reverse inner optimization. But the reverse inner
optimization works with prefixes, not suffixes. So the comparison wasn't
quite correct.

This goes back to only applying the reverse suffix optimization when
there is a non-empty single common suffix.

Fixes #1110
Ref https://github.com/astral-sh/ruff/pull/7980