cli: include \w, \s and \d in Unicode data table generation
This was an oversight omission when porting the old generator shell
script to regex-cli. This hasn't been an issue because I don't think
we've generated data for a new release of Unicode with this new
infrastructure yet.

This was flagged by unit tests that failed because \d was no longer a
subset of \w.