Reduce crate size by excluding fuzz/ and record/
The fuzz/ and record/ directories are about 2MiB, and are not used by
the end user. This removes them from the crate for future publishes.

Also, this avoids risk of an [xz-style attack] leveraging the binary
files produced by fuzzing, where malicious code could be slipped in past
review through binary test data.

[zx-style attack]: https://en.wikipedia.org/wiki/XZ_Utils_backdoor