ci: refactor testing
Recently, `cross test` started running doc tests. This is A Good Thing,
but the tests take *forever* because each test is compiled individually.

In Rust 2024, this should hopefully be fixed since most doc tests will
be combined together into one executable. But we haven't moved to Rust
2024 yet.

We do this by splitting up the tests for "native" targets versus "cross"
targets. This overall leads to a simpler configuration IMO.