api: tweak the lifetime of `Captures::get_match`
This was incredibly subtle and easy to miss. I actually didn't see any
warnings for this until I did `cargo publish` for the `1.12.0` release.
Not sure why.

I'll have to yank the `1.12.0` release because this is technically a
breaking change from the API published in `1.12.0`. Sigh. In particular,
in `1.12.0`, I assumed the inferred lifetime is tied to the borrow of
`Captures`, which is less than or equal to the lifetime of the haystack
`'h`.