docs: swap `doc_auto_cfg` with `doc_cfg`
In https://github.com/rust-lang/rust/pull/138907, the `doc_auto_cfg`
feature was subsumed by `doc_cfg`. This does overall look like we're on
a path toward stabilization, which is great.

One problem here though is that a bunch of crates use the `cfg` `docsrs`
to enable `doc_auto_cfg`. So if we enable it, then it causes those
crates to stop compiling. This is overall very annoying, and I don't
know how to unfuck things. So I changed the `cfg` knob to `docsrs_regex`
which, doesn't quite provide a guarantee, but gets us closer to being
masters of our own destiny.